refactor(perl-lsp): split lsp_comprehensive_3_17_test.rs into feature-specific test files
Split the monolithic 2,056-line, 60-test file into 15 focused test files
organized by LSP feature area:

- lsp_3_17_lifecycle_tests.rs (6 tests) - initialize, shutdown, error codes
- lsp_3_17_sync_tests.rs (1 test) - text document synchronization
- lsp_3_17_completion_tests.rs (1 test) - completion
- lsp_3_17_navigation_tests.rs (9 tests) - hover, definition, references, etc.
- lsp_3_17_code_actions_tests.rs (2 tests) - code actions and resolve
- lsp_3_17_document_features_tests.rs (6 tests) - code lens, links, folding, etc.
- lsp_3_17_formatting_tests.rs (3 tests) - formatting, range, on-type
- lsp_3_17_rename_tests.rs (2 tests) - rename and prepare rename
- lsp_3_17_semantic_tokens_tests.rs (2 tests) - semantic tokens
- lsp_3_17_hierarchy_tests.rs (4 tests) - call and type hierarchy
- lsp_3_17_diagnostics_tests.rs (5 tests) - diagnostics, inlay hints, moniker
- lsp_3_17_workspace_tests.rs (6 tests) - workspace symbols, commands, file ops
- lsp_3_17_window_tests.rs (9 tests) - window, progress, trace, telemetry
- lsp_3_17_notebook_tests.rs (2 tests) - notebook document support
- lsp_3_17_compliance_tests.rs (2 tests) - compliance validation
