--- crates/perl-lsp/tests/lsp_comprehensive_3_17_test.rs
@@ -1,2056 +1,0 @@
-//! Comprehensive LSP 3.17 API Contract Tests
-//!
-//! This test suite validates full compliance with the Language Server Protocol 3.17 specification.
-//! Every method, notification, and contract defined in the spec is tested here.
-//!
-
-#![allow(clippy::collapsible_if)]
-#![recursion_limit = "256"]
-//! Reference: https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/
-
-mod support;
-
-use serde_json::json;
-use support::lsp_harness::LspHarness;
-
-type TestResult = Result<(), Box<dyn std::error::Error>>;
-
-// ==================== LIFECYCLE CONTRACTS ====================
-
-#[test]
-fn test_initialize_contract_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-
-    // Full 3.17 initialization with all capabilities
-    let result = harness.initialize(Some(json!({
-        "processId": 1234,
-        "clientInfo": {
-            "name": "test-client",
-            "version": "1.0.0"
-        },
-        "locale": "en-US",
-        "rootPath": null,  // deprecated but still sent
-        "rootUri": "file:///workspace",
-        "capabilities": {
-            // 3.17 position encoding support
-            "general": {
-                "positionEncodings": ["utf-16", "utf-8", "utf-32"],
-                "staleRequestSupport": {
-                    "cancel": true,
-                    "retryOnContentModified": ["textDocument/completion"]
-                },
-                "regularExpressions": {
-                    "engine": "ECMAScript",
-                    "version": "ES2020"
-                },
-                "markdown": {
-                    "parser": "marked",
-                    "version": "1.0.0"
-                }
-            },
-            // Text document capabilities
-            "textDocument": {
-                "synchronization": {
-                    "dynamicRegistration": true,
-                    "willSave": true,
-                    "willSaveWaitUntil": true,
-                    "didSave": true
-                },
-                "completion": {
-                    "dynamicRegistration": true,
-                    "completionItem": {
-                        "snippetSupport": true,
-                        "commitCharactersSupport": true,
-                        "documentationFormat": ["markdown", "plaintext"],
-                        "deprecatedSupport": true,
-                        "preselectSupport": true,
-                        "tagSupport": { "valueSet": [1] },
-                        "insertReplaceSupport": true,
-                        "resolveSupport": {
-                            "properties": ["documentation", "detail", "additionalTextEdits"]
-                        },
-                        "insertTextModeSupport": { "valueSet": [1, 2] },
-                        "labelDetailsSupport": true
-                    },
-                    "completionItemKind": { "valueSet": [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25] },
-                    "insertTextMode": 2,
-                    "contextSupport": true,
-                    "completionList": {
-                        "itemDefaults": ["commitCharacters", "editRange", "insertTextFormat", "insertTextMode", "data"]
-                    }
-                },
-                "hover": {
-                    "dynamicRegistration": true,
-                    "contentFormat": ["markdown", "plaintext"]
-                },
-                "signatureHelp": {
-                    "dynamicRegistration": true,
-                    "signatureInformation": {
-                        "documentationFormat": ["markdown", "plaintext"],
-                        "parameterInformation": { "labelOffsetSupport": true },
-                        "activeParameterSupport": true
-                    },
-                    "contextSupport": true
-                },
-                "declaration": {
-                    "dynamicRegistration": true,
-                    "linkSupport": true
-                },
-                "definition": {
-                    "dynamicRegistration": true,
-                    "linkSupport": true
-                },
-                "typeDefinition": {
-                    "dynamicRegistration": true,
-                    "linkSupport": true
-                },
-                "implementation": {
-                    "dynamicRegistration": true,
-                    "linkSupport": true
-                },
-                "references": {
-                    "dynamicRegistration": true
-                },
-                "documentHighlight": {
-                    "dynamicRegistration": true
-                },
-                "documentSymbol": {
-                    "dynamicRegistration": true,
-                    "symbolKind": { "valueSet": [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26] },
-                    "hierarchicalDocumentSymbolSupport": true,
-                    "tagSupport": { "valueSet": [1] },
-                    "labelSupport": true
-                },
-                "codeAction": {
-                    "dynamicRegistration": true,
-                    "codeActionLiteralSupport": {
-                        "codeActionKind": {
-                            "valueSet": ["", "quickfix", "refactor", "refactor.extract", "refactor.inline", "refactor.rewrite", "source", "source.organizeImports", "source.fixAll"]
-                        }
-                    },
-                    "isPreferredSupport": true,
-                    "disabledSupport": true,
-                    "dataSupport": true,
-                    "resolveSupport": {
-                        "properties": ["edit"]
-                    },
-                    "honorsChangeAnnotations": true
-                },
-                "codeLens": {
-                    "dynamicRegistration": true
-                },
-                "documentLink": {
-                    "dynamicRegistration": true,
-                    "tooltipSupport": true
-                },
-                "colorProvider": {
-                    "dynamicRegistration": true
-                },
-                "formatting": {
-                    "dynamicRegistration": true
-                },
-                "rangeFormatting": {
-                    "dynamicRegistration": true,
-                    "rangesSupport": true
-                },
-                "onTypeFormatting": {
-                    "dynamicRegistration": true
-                },
-                "rename": {
-                    "dynamicRegistration": true,
-                    "prepareSupport": true,
-                    "prepareSupportDefaultBehavior": 1,
-                    "honorsChangeAnnotations": true
-                },
-                "foldingRange": {
-                    "dynamicRegistration": true,
-                    "rangeLimit": 5000,
-                    "lineFoldingOnly": false,
-                    "foldingRangeKind": { "valueSet": ["comment", "imports", "region"] },
-                    "foldingRange": { "collapsedText": false }
-                },
-                "selectionRange": {
-                    "dynamicRegistration": true
-                },
-                "publishDiagnostics": {
-                    "relatedInformation": true,
-                    "tagSupport": { "valueSet": [1, 2] },
-                    "versionSupport": true,
-                    "codeDescriptionSupport": true,
-                    "dataSupport": true
-                },
-                "callHierarchy": {
-                    "dynamicRegistration": true
-                },
-                "semanticTokens": {
-                    "dynamicRegistration": true,
-                    "requests": {
-                        "range": true,
-                        "full": { "delta": true }
-                    },
-                    "tokenTypes": ["namespace", "type", "class", "enum", "interface", "struct", "typeParameter", "parameter", "variable", "property", "enumMember", "event", "function", "method", "macro", "keyword", "modifier", "comment", "string", "number", "regexp", "operator", "decorator"],
-                    "tokenModifiers": ["declaration", "definition", "readonly", "static", "deprecated", "abstract", "async", "modification", "documentation", "defaultLibrary"],
-                    "formats": ["relative"],
-                    "overlappingTokenSupport": false,
-                    "multilineTokenSupport": true,
-                    "serverCancelSupport": true,
-                    "augmentsSyntaxTokens": true
-                },
-                "linkedEditingRange": {
-                    "dynamicRegistration": true
-                },
-                "typeHierarchy": {
-                    "dynamicRegistration": true
-                },
-                "inlineValue": {
-                    "dynamicRegistration": true
-                },
-                "inlayHint": {
-                    "dynamicRegistration": true,
-                    "resolveSupport": {
-                        "properties": ["tooltip", "textEdits", "label.tooltip", "label.location", "label.command"]
-                    }
-                },
-                "diagnostic": {
-                    "dynamicRegistration": true,
-                    "relatedDocumentSupport": true
-                },
-                "moniker": {
-                    "dynamicRegistration": true
-                }
-            },
-            // Notebook document support (3.17)
-            "notebookDocument": {
-                "synchronization": {
-                    "dynamicRegistration": true,
-                    "executionSummarySupport": true
-                }
-            },
-            // Window capabilities
-            "window": {
-                "workDoneProgress": true,
-                "showMessage": {
-                    "messageActionItem": {
-                        "additionalPropertiesSupport": true
-                    }
-                },
-                "showDocument": {
-                    "support": true
-                }
-            },
-            // Workspace capabilities
-            "workspace": {
-                "applyEdit": true,
-                "workspaceEdit": {
-                    "documentChanges": true,
-                    "resourceOperations": ["create", "rename", "delete"],
-                    "failureHandling": "textOnlyTransactional",
-                    "normalizesLineEndings": true,
-                    "changeAnnotationSupport": {
-                        "groupsOnLabel": true
-                    }
-                },
-                "didChangeConfiguration": {
-                    "dynamicRegistration": true
-                },
-                "didChangeWatchedFiles": {
-                    "dynamicRegistration": true,
-                    "relativePatternSupport": true
-                },
-                "symbol": {
-                    "dynamicRegistration": true,
-                    "symbolKind": { "valueSet": [1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26] },
-                    "tagSupport": { "valueSet": [1] },
-                    "resolveSupport": {
-                        "properties": ["location.range"]
-                    }
-                },
-                "executeCommand": {
-                    "dynamicRegistration": true
-                },
-                "semanticTokens": {
-                    "refreshSupport": true
-                },
-                "codeLens": {
-                    "refreshSupport": true
-                },
-                "fileOperations": {
-                    "dynamicRegistration": true,
-                    "didCreate": true,
-                    "willCreate": true,
-                    "didRename": true,
-                    "willRename": true,
-                    "didDelete": true,
-                    "willDelete": true
-                },
-                "inlineValue": {
-                    "refreshSupport": true
-                },
-                "inlayHint": {
-                    "refreshSupport": true
-                },
-                "diagnostics": {
-                    "refreshSupport": true
-                },
-                "workspaceFolders": true,
-                "configuration": true
-            }
-        },
-        "initializationOptions": {
-            "testMode": true
-        },
-        "workspaceFolders": [
-            {
-                "uri": "file:///workspace",
-                "name": "Test Workspace"
-            }
-        ]
-    })))?;
-
-    // Validate server response structure
-    assert!(result.is_object());
-    let capabilities = &result["capabilities"];
-    assert!(capabilities.is_object());
-
-    // Check position encoding (3.17)
-    if let Some(encoding) = capabilities.get("positionEncoding") {
-        assert!(encoding.is_string());
-        let enc = encoding.as_str().ok_or("encoding not a string")?;
-        assert!(["utf-8", "utf-16", "utf-32"].contains(&enc));
-    }
-
-    // Check server info
-    if let Some(info) = result.get("serverInfo") {
-        assert!(info["name"].is_string());
-        if let Some(version) = info.get("version") {
-            assert!(version.is_string());
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_initialized_notification() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Send initialized notification - no response expected
-    harness.notify("initialized", json!({}));
-
-    // Server should now accept requests
-    let response = harness.request(
-        "textDocument/hover",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 0 }
-        }),
-    );
-
-    // Should get a valid response (even if null)
-    assert!(response.is_ok() || response.is_err());
-    Ok(())
-}
-
-// ==================== TEXT SYNCHRONIZATION ====================
-
-#[test]
-fn test_text_document_sync_incremental() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // didOpen
-    harness.notify(
-        "textDocument/didOpen",
-        json!({
-            "textDocument": {
-                "uri": "file:///test.pl",
-                "languageId": "perl",
-                "version": 1,
-                "text": "my $x = 42;\n"
-            }
-        }),
-    );
-
-    // didChange (full content — still valid under incremental sync)
-    harness.notify(
-        "textDocument/didChange",
-        json!({
-            "textDocument": {
-                "uri": "file:///test.pl",
-                "version": 2
-            },
-            "contentChanges": [
-                { "text": "my $x = 43;\nmy $y = $x;\n" }
-            ]
-        }),
-    );
-
-    // didChange (incremental / range-based)
-    harness.notify(
-        "textDocument/didChange",
-        json!({
-            "textDocument": {
-                "uri": "file:///test.pl",
-                "version": 3
-            },
-            "contentChanges": [
-                {
-                    "range": {
-                        "start": { "line": 0, "character": 9 },
-                        "end": { "line": 0, "character": 11 }
-                    },
-                    "text": "99"
-                }
-            ]
-        }),
-    );
-
-    // willSave
-    harness.notify(
-        "textDocument/willSave",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "reason": 1  // Manual
-        }),
-    );
-
-    // willSaveWaitUntil - expects response
-    let edits = harness.request(
-        "textDocument/willSaveWaitUntil",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "reason": 1
-        }),
-    );
-
-    if let Ok(edits) = edits {
-        assert!(edits.is_array() || edits.is_null());
-    }
-
-    // didSave
-    harness.notify(
-        "textDocument/didSave",
-        json!({
-            "textDocument": { "uri": "file:///test.pl", "version": 4 },
-            "text": "my $x = 43;\nmy $y = $x;\n"  // optional
-        }),
-    );
-
-    // didClose
-    harness.notify(
-        "textDocument/didClose",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    );
-    Ok(())
-}
-
-// ==================== LANGUAGE FEATURES ====================
-
-#[test]
-fn test_completion_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "print $")?;
-
-    let response = harness.request(
-        "textDocument/completion",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 7 },
-            "context": {
-                "triggerKind": 1,  // Invoked
-                "triggerCharacter": "$"
-            }
-        }),
-    )?;
-
-    // Response can be array or CompletionList
-    assert!(response.is_array() || (response.is_object() && response.get("items").is_some()));
-    Ok(())
-}
-
-#[test]
-fn test_hover_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "print 'hello'")?;
-
-    let response = harness.request(
-        "textDocument/hover",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 0 },
-            "workDoneToken": "hover-1"  // optional progress token
-        }),
-    )?;
-
-    if !response.is_null() {
-        assert!(
-            response["contents"].is_string()
-                || response["contents"].is_object()
-                || response["contents"].is_array()
-        );
-    }
-    Ok(())
-}
-
-#[test]
-fn test_signature_help_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "substr(")?;
-
-    let response = harness.request(
-        "textDocument/signatureHelp",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 7 },
-            "context": {
-                "triggerKind": 1,  // Invoked
-                "triggerCharacter": "(",
-                "isRetrigger": false,
-                "activeSignatureHelp": null
-            }
-        }),
-    )?;
-
-    if !response.is_null() {
-        assert!(response["signatures"].is_array());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_declaration_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $x = 1;\n$x")?;
-
-    let response = harness.request(
-        "textDocument/declaration",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 1, "character": 0 }
-        }),
-    )?;
-
-    // Can be Location, Location[], LocationLink[], or null
-    assert!(response.is_null() || response.is_object() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_definition_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "sub test {}\ntest()")?;
-
-    let response = harness.request(
-        "textDocument/definition",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 1, "character": 0 }
-        }),
-    )?;
-
-    // Can be Location, Location[], LocationLink[], or null
-    assert!(response.is_null() || response.is_object() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_type_definition_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    let init_response = harness.initialize(None)?;
-    let caps = &init_response["capabilities"];
-
-    // Check if server advertises typeDefinition support
-    let supported =
-        caps.get("typeDefinitionProvider").is_some() && !caps["typeDefinitionProvider"].is_null();
-
-    harness.open("file:///test.pl", "my $obj = bless {}, 'MyClass'")?;
-
-    let response = harness.request_raw(json!({
-        "jsonrpc": "2.0",
-        "id": 1,
-        "method": "textDocument/typeDefinition",
-        "params": {
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 }
-        }
-    }));
-
-    if supported {
-        // If supported, should return a result (which could be an array or locations)
-        // The test harness returns the result directly, not wrapped in a response
-        assert!(
-            response.is_array() || response.is_object() || response.is_null(),
-            "Expected array, object, or null result for typeDefinition"
-        );
-    } else {
-        // If not supported, should return an error (MethodNotFound or InternalError)
-        assert!(response.get("error").is_some(), "Expected error when not advertised");
-        let error_code = response["error"]["code"].as_i64().ok_or("error code not i64")?;
-        assert!(
-            error_code == -32601 || error_code == -32603,
-            "Expected MethodNotFound (-32601) or InternalError (-32603), got {}",
-            error_code
-        );
-    }
-    Ok(())
-}
-
-#[test]
-fn test_implementation_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    let init_response = harness.initialize(None)?;
-    let caps = &init_response["capabilities"];
-
-    // Check if server advertises implementation support
-    let supported =
-        caps.get("implementationProvider").is_some() && !caps["implementationProvider"].is_null();
-
-    harness.open(
-        "file:///test.pl",
-        "package Base;\nsub method {}\npackage Derived;\nuse base 'Base';",
-    )?;
-
-    let response = harness.request_raw(json!({
-        "jsonrpc": "2.0",
-        "id": 1,
-        "method": "textDocument/implementation",
-        "params": {
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 1, "character": 4 }
-        }
-    }));
-
-    if supported {
-        // If supported, should return a result (which could be an array or locations)
-        // The test harness returns the result directly, not wrapped in a response
-        assert!(
-            response.is_array() || response.is_object(),
-            "Expected array or object result for implementation"
-        );
-    } else {
-        // If not supported, should return an error (MethodNotFound or InternalError)
-        assert!(response.get("error").is_some(), "Expected error when not advertised");
-        let error_code = response["error"]["code"].as_i64().ok_or("error code not i64")?;
-        assert!(
-            error_code == -32601 || error_code == -32603,
-            "Expected MethodNotFound (-32601) or InternalError (-32603), got {}",
-            error_code
-        );
-    }
-    Ok(())
-}
-
-#[test]
-fn test_references_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $x = 1;\n$x++;\nprint $x;")?;
-
-    let response = harness.request(
-        "textDocument/references",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 },
-            "context": {
-                "includeDeclaration": true
-            }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_document_highlight_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $x = 1;\n$x = 2;\nprint $x;")?;
-
-    let response = harness.request(
-        "textDocument/documentHighlight",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_document_symbol_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "package Foo;\nsub bar {}\nmy $var = 1;")?;
-
-    let response = harness.request(
-        "textDocument/documentSymbol",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_code_action_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "$undefined")?;
-
-    let response = harness.request(
-        "textDocument/codeAction",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "range": {
-                "start": { "line": 0, "character": 0 },
-                "end": { "line": 0, "character": 10 }
-            },
-            "context": {
-                "diagnostics": [],
-                "only": ["quickfix", "refactor"],
-                "triggerKind": 1  // Invoked
-            }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_code_action_resolve_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Mock code action to resolve
-    let response = harness.request(
-        "codeAction/resolve",
-        json!({
-            "title": "Extract variable",
-            "kind": "refactor.extract",
-            "data": { "uri": "file:///test.pl", "range": {} }
-        }),
-    );
-
-    // May fail if not supported
-    if let Ok(action) = response {
-        assert!(action["title"].is_string());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_code_lens_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "sub test {}\ntest();")?;
-
-    let response = harness.request(
-        "textDocument/codeLens",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_document_link_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "use strict;\nuse Data::Dumper;")?;
-
-    let response = harness.request(
-        "textDocument/documentLink",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_document_color_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.css", ".class { color: #FF0000; }")?;
-
-    // May not be supported for Perl
-    let response = harness.request(
-        "textDocument/documentColor",
-        json!({
-            "textDocument": { "uri": "file:///test.css" }
-        }),
-    );
-
-    if let Ok(colors) = response {
-        assert!(colors.is_null() || colors.is_array());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_formatting_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my$x=1;print$x;")?;
-
-    let response = harness.request(
-        "textDocument/formatting",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "options": {
-                "tabSize": 4,
-                "insertSpaces": true,
-                "trimTrailingWhitespace": true,
-                "insertFinalNewline": true,
-                "trimFinalNewlines": true
-            }
-        }),
-    );
-
-    // Handle both success and error cases - this is a protocol compliance test
-    match response {
-        Ok(result) => {
-            // Success: should return null or array of edits
-            assert!(result.is_null() || result.is_array());
-        }
-        Err(_) => {
-            // Error is acceptable when perltidy is not available
-            // This maintains LSP protocol compliance
-            eprintln!(
-                "Formatting failed (perltidy may not be installed) - this is acceptable for protocol compliance"
-            );
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_range_formatting_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my$x=1;\nprint$x;")?;
-
-    let response = harness.request(
-        "textDocument/rangeFormatting",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "range": {
-                "start": { "line": 0, "character": 0 },
-                "end": { "line": 0, "character": 7 }
-            },
-            "options": {
-                "tabSize": 4,
-                "insertSpaces": true
-            }
-        }),
-    );
-
-    // Handle both success and error cases - this is a protocol compliance test
-    match response {
-        Ok(result) => {
-            // Success: should return null or array of edits
-            assert!(result.is_null() || result.is_array());
-        }
-        Err(_) => {
-            // Error is acceptable when perltidy is not available
-            // This maintains LSP protocol compliance
-            eprintln!(
-                "Range formatting failed (perltidy may not be installed) - this is acceptable for protocol compliance"
-            );
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_on_type_formatting_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "if (1) {")?;
-
-    let response = harness.request(
-        "textDocument/onTypeFormatting",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 8 },
-            "ch": "{",
-            "options": {
-                "tabSize": 4,
-                "insertSpaces": true
-            }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_rename_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $old = 1;\n$old++;")?;
-
-    let response = harness.request(
-        "textDocument/rename",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 },
-            "newName": "new"
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_object());
-    Ok(())
-}
-
-#[test]
-fn test_prepare_rename_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $var = 1;")?;
-
-    let response = harness.request(
-        "textDocument/prepareRename",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 }
-        }),
-    )?;
-
-    // Can be Range, {range, placeholder}, {defaultBehavior}, or null
-    assert!(response.is_null() || response.is_object() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_folding_range_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "sub test {\n    my $x = 1;\n    return $x;\n}")?;
-
-    let response = harness.request(
-        "textDocument/foldingRange",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_selection_range_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "if ($x) { print $x; }")?;
-
-    let response = harness.request(
-        "textDocument/selectionRange",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "positions": [
-                { "line": 0, "character": 10 }
-            ]
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_linked_editing_range_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "<div></div>")?;
-
-    let response = harness.request(
-        "textDocument/linkedEditingRange",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 1 }
-        }),
-    );
-
-    if let Ok(ranges) = response {
-        assert!(ranges.is_null() || (ranges.is_object() && ranges["ranges"].is_array()));
-    }
-    Ok(())
-}
-
-// ==================== SEMANTIC TOKENS (3.16+) ====================
-
-#[test]
-fn test_semantic_tokens_full_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "package Foo;\nsub bar { my $var = 1; }")?;
-
-    let response = harness.request(
-        "textDocument/semanticTokens/full",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" }
-        }),
-    );
-
-    if let Ok(tokens) = response {
-        if !tokens.is_null() {
-            assert!(tokens["data"].is_array());
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_semantic_tokens_range_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $x = 1;\nmy $y = 2;")?;
-
-    let response = harness.request(
-        "textDocument/semanticTokens/range",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "range": {
-                "start": { "line": 0, "character": 0 },
-                "end": { "line": 0, "character": 10 }
-            }
-        }),
-    );
-
-    if let Ok(tokens) = response {
-        assert!(tokens.is_null() || tokens["data"].is_array());
-    }
-    Ok(())
-}
-
-// ==================== CALL HIERARCHY (3.16+) ====================
-
-#[test]
-fn test_prepare_call_hierarchy_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "sub test { helper(); }\nsub helper {}")?;
-
-    let response = harness.request(
-        "textDocument/prepareCallHierarchy",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 0, "character": 4 }
-        }),
-    );
-
-    if let Ok(items) = response {
-        assert!(items.is_null() || items.is_array());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_incoming_calls_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    let response = harness.request(
-        "callHierarchy/incomingCalls",
-        json!({
-            "item": {
-                "name": "test",
-                "kind": 12,  // Function
-                "uri": "file:///test.pl",
-                "range": {
-                    "start": { "line": 0, "character": 0 },
-                    "end": { "line": 0, "character": 20 }
-                },
-                "selectionRange": {
-                    "start": { "line": 0, "character": 4 },
-                    "end": { "line": 0, "character": 8 }
-                }
-            }
-        }),
-    );
-
-    if let Ok(calls) = response {
-        assert!(calls.is_null() || calls.is_array());
-    }
-    Ok(())
-}
-
-// ==================== TYPE HIERARCHY (3.17) ====================
-
-#[test]
-fn test_prepare_type_hierarchy_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "package Base;\npackage Derived;\nuse base 'Base';")?;
-
-    let response = harness.request(
-        "textDocument/prepareTypeHierarchy",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 1, "character": 8 }
-        }),
-    );
-
-    if let Ok(items) = response {
-        assert!(items.is_null() || items.is_array());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_type_hierarchy_supertypes_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    let response = harness.request(
-        "typeHierarchy/supertypes",
-        json!({
-            "item": {
-                "name": "Derived",
-                "kind": 5,  // Class
-                "uri": "file:///test.pl",
-                "range": {
-                    "start": { "line": 1, "character": 0 },
-                    "end": { "line": 2, "character": 17 }
-                },
-                "selectionRange": {
-                    "start": { "line": 1, "character": 8 },
-                    "end": { "line": 1, "character": 15 }
-                }
-            }
-        }),
-    );
-
-    if let Ok(types) = response {
-        assert!(types.is_null() || types.is_array());
-    }
-    Ok(())
-}
-
-// ==================== INLAY HINTS (3.17) ====================
-
-#[test]
-fn test_inlay_hint_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "substr($str, 0, 5)")?;
-
-    let response = harness.request(
-        "textDocument/inlayHint",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "range": {
-                "start": { "line": 0, "character": 0 },
-                "end": { "line": 0, "character": 18 }
-            }
-        }),
-    );
-
-    if let Ok(hints) = response {
-        assert!(hints.is_null() || hints.is_array());
-    }
-    Ok(())
-}
-
-// ==================== INLINE VALUES (3.17) ====================
-
-#[test]
-fn test_inline_value_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "my $x = 42;\nprint $x;")?;
-
-    let response = harness.request(
-        "textDocument/inlineValue",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "range": {
-                "start": { "line": 0, "character": 0 },
-                "end": { "line": 1, "character": 9 }
-            },
-            "context": {
-                "frameId": 1,
-                "stoppedLocation": {
-                    "start": { "line": 1, "character": 0 },
-                    "end": { "line": 1, "character": 9 }
-                }
-            }
-        }),
-    );
-
-    if let Ok(values) = response {
-        assert!(values.is_null() || values.is_array());
-    }
-    Ok(())
-}
-
-// ==================== MONIKER (3.16+) ====================
-
-#[test]
-fn test_moniker_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "package Foo::Bar;\nsub test {}")?;
-
-    let response = harness.request(
-        "textDocument/moniker",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "position": { "line": 1, "character": 4 }
-        }),
-    );
-
-    if let Ok(monikers) = response {
-        assert!(monikers.is_null() || monikers.is_array());
-    }
-    Ok(())
-}
-
-// ==================== DIAGNOSTICS PULL MODEL (3.17) ====================
-
-#[test]
-fn test_diagnostic_pull_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "$undefined")?;
-
-    let response = harness.request(
-        "textDocument/diagnostic",
-        json!({
-            "textDocument": { "uri": "file:///test.pl" },
-            "identifier": "perl-lsp",
-            "previousResultId": null
-        }),
-    );
-
-    if let Ok(report) = response {
-        if !report.is_null() {
-            assert!(report["kind"].is_string());
-            if report["kind"] == "full" {
-                assert!(report["items"].is_array());
-            }
-        }
-    }
-    Ok(())
-}
-
-#[test]
-fn test_workspace_diagnostic_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    let response = harness.request(
-        "workspace/diagnostic",
-        json!({
-            "identifier": "perl-lsp",
-            "previousResultIds": [],
-            "workDoneToken": "diag-1",
-            "partialResultToken": "partial-1"
-        }),
-    );
-
-    if let Ok(report) = response {
-        assert!(report.is_null() || report.is_object());
-    }
-    Ok(())
-}
-
-// ==================== WORKSPACE FEATURES ====================
-
-#[test]
-fn test_workspace_symbol_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    let response = harness.request(
-        "workspace/symbol",
-        json!({
-            "query": "test",
-            "workDoneToken": "symbol-1"
-        }),
-    )?;
-
-    assert!(response.is_null() || response.is_array());
-    Ok(())
-}
-
-#[test]
-fn test_workspace_symbol_resolve_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Mock workspace symbol to resolve
-    let response = harness.request(
-        "workspaceSymbol/resolve",
-        json!({
-            "name": "test",
-            "kind": 12,
-            "location": {
-                "uri": "file:///test.pl"
-            }
-        }),
-    );
-
-    if let Ok(symbol) = response {
-        assert!(symbol["name"].is_string());
-    }
-    Ok(())
-}
-
-#[test]
-fn test_execute_command_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    let response = harness.request(
-        "workspace/executeCommand",
-        json!({
-            "command": "perl.extractVariable",
-            "arguments": [
-                "file:///test.pl",
-                { "start": { "line": 0, "character": 0 }, "end": { "line": 0, "character": 5 } }
-            ],
-            "workDoneToken": "cmd-1"
-        }),
-    );
-
-    // May fail if command not supported
-    if let Ok(_result) = response {
-        // Result can be any value
-    }
-    Ok(())
-}
-
-#[test]
-fn test_workspace_folders_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Notify of workspace folder changes
-    harness.notify(
-        "workspace/didChangeWorkspaceFolders",
-        json!({
-            "event": {
-                "added": [
-                    { "uri": "file:///workspace2", "name": "Workspace 2" }
-                ],
-                "removed": []
-            }
-        }),
-    );
-
-    // Server can also request current folders (if needed)
-    // This would be a server->client request, so we skip it in tests
-    Ok(())
-}
-
-#[test]
-fn test_file_operations_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // willCreateFiles
-    let response = harness.request(
-        "workspace/willCreateFiles",
-        json!({
-            "files": [
-                { "uri": "file:///new.pl" }
-            ]
-        }),
-    );
-
-    if let Ok(edit) = response {
-        assert!(edit.is_null() || edit.is_object());
-    }
-
-    // didCreateFiles
-    harness.notify(
-        "workspace/didCreateFiles",
-        json!({
-            "files": [
-                { "uri": "file:///new.pl" }
-            ]
-        }),
-    );
-
-    // willRenameFiles
-    let response = harness.request(
-        "workspace/willRenameFiles",
-        json!({
-            "files": [
-                { "oldUri": "file:///old.pl", "newUri": "file:///new.pl" }
-            ]
-        }),
-    );
-
-    if let Ok(edit) = response {
-        assert!(edit.is_null() || edit.is_object());
-    }
-
-    // didRenameFiles
-    harness.notify(
-        "workspace/didRenameFiles",
-        json!({
-            "files": [
-                { "oldUri": "file:///old.pl", "newUri": "file:///new.pl" }
-            ]
-        }),
-    );
-
-    // willDeleteFiles
-    let response = harness.request(
-        "workspace/willDeleteFiles",
-        json!({
-            "files": [
-                { "uri": "file:///delete.pl" }
-            ]
-        }),
-    );
-
-    if let Ok(edit) = response {
-        assert!(edit.is_null() || edit.is_object());
-    }
-
-    // didDeleteFiles
-    harness.notify(
-        "workspace/didDeleteFiles",
-        json!({
-            "files": [
-                { "uri": "file:///delete.pl" }
-            ]
-        }),
-    );
-    Ok(())
-}
-
-#[test]
-fn test_watched_files_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    harness.notify(
-        "workspace/didChangeWatchedFiles",
-        json!({
-            "changes": [
-                { "uri": "file:///test.pl", "type": 2 },  // Changed
-                { "uri": "file:///new.pl", "type": 1 },   // Created
-                { "uri": "file:///old.pl", "type": 3 }    // Deleted
-            ]
-        }),
-    );
-    Ok(())
-}
-
-// ==================== WINDOW FEATURES ====================
-
-#[test]
-fn test_show_message_3_17() -> TestResult {
-    // Server->client, so we can't test directly
-    // But we document the contract here
-
-    // window/showMessage notification
-    // { "type": 1-4, "message": "string" }
-
-    // window/showMessageRequest
-    // { "type": 1-4, "message": "string", "actions": [{"title": "string"}] }
-    // Response: MessageActionItem | null
-    Ok(())
-}
-
-#[test]
-fn test_show_document_3_17() -> TestResult {
-    // Server->client request (3.16+)
-    // Params: { uri, external?, takeFocus?, selection? }
-    // Response: { success: boolean }
-    Ok(())
-}
-
-#[test]
-fn test_log_message_3_17() -> TestResult {
-    // Server->client notification
-    // { "type": 1-4, "message": "string" }
-    Ok(())
-}
-
-#[test]
-fn test_work_done_progress_3_17() -> TestResult {
-    // window/workDoneProgress/create (server->client)
-    // $/progress notifications
-    // window/workDoneProgress/cancel (client->server)
-    Ok(())
-}
-
-// ==================== MISCELLANEOUS ====================
-
-#[test]
-fn test_cancel_request_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Send a request then immediately cancel it
-    // In real scenario, this would be async
-    harness.notify(
-        "$/cancelRequest",
-        json!({
-            "id": 999
-        }),
-    );
-    Ok(())
-}
-
-#[test]
-fn test_progress_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Progress notifications can flow either way
-    harness.notify(
-        "$/progress",
-        json!({
-            "token": "test-progress",
-            "value": {
-                "kind": "begin",
-                "title": "Processing",
-                "cancellable": true,
-                "percentage": 0
-            }
-        }),
-    );
-
-    harness.notify(
-        "$/progress",
-        json!({
-            "token": "test-progress",
-            "value": {
-                "kind": "report",
-                "message": "Working...",
-                "percentage": 50
-            }
-        }),
-    );
-
-    harness.notify(
-        "$/progress",
-        json!({
-            "token": "test-progress",
-            "value": {
-                "kind": "end",
-                "message": "Complete"
-            }
-        }),
-    );
-    Ok(())
-}
-
-#[test]
-fn test_set_trace_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    harness.notify(
-        "$/setTrace",
-        json!({
-            "value": "verbose"  // off | messages | verbose
-        }),
-    );
-    Ok(())
-}
-
-#[test]
-fn test_log_trace_3_17() -> TestResult {
-    // Server->client notification when tracing is on
-    // { "message": "string", "verbose"?: "string" }
-    Ok(())
-}
-
-#[test]
-fn test_telemetry_3_17() -> TestResult {
-    // Server->client notification
-    // params: any
-    Ok(())
-}
-
-// ==================== SHUTDOWN & EXIT ====================
-
-#[test]
-fn test_shutdown_exit_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Shutdown request
-    let response = harness.request("shutdown", json!(null))?;
-    assert!(response.is_null());
-
-    // Exit notification
-    harness.notify("exit", json!(null));
-    Ok(())
-}
-
-// ==================== ERROR CODES ====================
-
-#[test]
-fn test_error_codes_3_17() -> TestResult {
-    // Standard JSON-RPC error codes
-    const _PARSE_ERROR: i32 = -32700;
-    #[allow(dead_code)]
-    const INVALID_REQUEST: i32 = -32600;
-    #[allow(dead_code)]
-    const METHOD_NOT_FOUND: i32 = -32601;
-    #[allow(dead_code)]
-    const INVALID_PARAMS: i32 = -32602;
-    #[allow(dead_code)]
-    const INTERNAL_ERROR: i32 = -32603;
-
-    // LSP error codes
-    const _SERVER_NOT_INITIALIZED: i32 = -32002;
-    #[allow(dead_code)]
-    const UNKNOWN_ERROR_CODE: i32 = -32001;
-    const _REQUEST_CANCELLED: i32 = -32800;
-    #[allow(dead_code)]
-    const CONTENT_MODIFIED: i32 = -32801;
-    const _SERVER_CANCELLED: i32 = -32802; // 3.17
-    const _REQUEST_FAILED: i32 = -32803;
-
-    // Validate error code constants match spec
-    // PARSE_ERROR = -32700 (< -32000 as required)
-    // SERVER_NOT_INITIALIZED = -32002
-    // REQUEST_CANCELLED = -32800
-    // SERVER_CANCELLED = -32802 (LSP 3.17)
-    // REQUEST_FAILED = -32803
-    Ok(())
-}
-
-// ==================== PRE-INITIALIZE BEHAVIOR ====================
-
-#[test]
-fn test_inbound_before_initialize_contract() -> TestResult {
-    // Requests before initialize must return -32002 ServerNotInitialized
-    // Notifications must be dropped (except exit)
-
-    // This test would need a harness method to create without auto-initialize
-    // let mut harness = LspHarness::new_without_initialize();
-
-    // Request before initialize -> -32002
-    // let resp = harness.request_raw(json!({
-    //     "jsonrpc":"2.0","id":1,"method":"textDocument/hover",
-    //     "params":{"textDocument":{"uri":"file:///t.pl"},
-    //               "position":{"line":0,"character":0}}
-    // }));
-    // assert_eq!(resp["error"]["code"], -32002);
-
-    // Notification before initialize -> drop silently
-    // harness.notify("workspace/didChangeConfiguration", json!({"settings":{}}));
-    Ok(())
-}
-
-// ==================== $-PREFIXED MESSAGES ====================
-
-#[test]
-fn test_dollar_prefixed_request_method_not_found() -> TestResult {
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-
-    // Requests with methods starting with $/ must return -32601 MethodNotFound
-    // (unless explicitly implemented like $/cancelRequest)
-
-    // This would test unknown $/ methods
-    // let resp = harness.request_raw(json!({
-    //     "jsonrpc":"2.0","id":1,"method":"$/unknownRequest","params":{}
-    // }));
-    // assert_eq!(resp["error"]["code"], -32601);
-    Ok(())
-}
-
-// ==================== NOTEBOOK SUPPORT (3.17) ====================
-
-#[test]
-fn test_notebook_document_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    let init = harness.initialize(None)?;
-    assert!(
-        init["capabilities"]["notebookDocumentSync"].is_object(),
-        "notebookDocumentSync capability should be advertised"
-    );
-
-    // didOpen notebook
-    harness.notify(
-        "notebookDocument/didOpen",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///test.ipynb",
-                "notebookType": "jupyter-notebook",
-                "version": 1,
-                "cells": [
-                    {
-                        "kind": 2,  // Code
-                        "document": "file:///test.ipynb#cell1"
-                    }
-                ]
-            },
-            "cellTextDocuments": [
-                {
-                    "uri": "file:///test.ipynb#cell1",
-                    "languageId": "perl",
-                    "version": 1,
-                    "text": "sub from_notebook_cell { return 1; }\n"
-                }
-            ]
-        }),
-    );
-
-    let cell1_symbols = harness.request(
-        "textDocument/documentSymbol",
-        json!({
-            "textDocument": {
-                "uri": "file:///test.ipynb#cell1"
-            }
-        }),
-    )?;
-    let cell1_symbols = cell1_symbols.as_array().ok_or("cell1 symbols should be an array")?;
-    assert!(
-        cell1_symbols.iter().any(|symbol| symbol["name"].as_str() == Some("from_notebook_cell")),
-        "Expected symbol from_notebook_cell in notebook cell document symbols"
-    );
-
-    // didChange notebook
-    harness.notify(
-        "notebookDocument/didChange",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///test.ipynb",
-                "version": 2
-            },
-            "change": {
-                "cells": {
-                    "structure": {
-                        "array": {
-                            "start": 0,
-                            "deleteCount": 0,
-                            "cells": [
-                                {
-                                    "kind": 2,
-                                    "document": "file:///test.ipynb#cell2"
-                                }
-                            ]
-                        },
-                        "didOpen": [
-                            {
-                                "uri": "file:///test.ipynb#cell2",
-                                "languageId": "perl",
-                                "version": 1,
-                                "text": "sub second_notebook_cell { return 42; }\n"
-                            }
-                        ]
-                    }
-                }
-            }
-        }),
-    );
-
-    let cell2_symbols = harness.request(
-        "textDocument/documentSymbol",
-        json!({
-            "textDocument": {
-                "uri": "file:///test.ipynb#cell2"
-            }
-        }),
-    )?;
-    let cell2_symbols = cell2_symbols.as_array().ok_or("cell2 symbols should be an array")?;
-    assert!(
-        cell2_symbols.iter().any(|symbol| symbol["name"].as_str() == Some("second_notebook_cell")),
-        "Expected symbol second_notebook_cell in newly opened notebook cell"
-    );
-
-    // didSave notebook
-    harness.notify(
-        "notebookDocument/didSave",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///test.ipynb"
-            }
-        }),
-    );
-
-    // didClose notebook
-    harness.notify(
-        "notebookDocument/didClose",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///test.ipynb"
-            },
-            "cellTextDocuments": [
-                { "uri": "file:///test.ipynb#cell1" },
-                { "uri": "file:///test.ipynb#cell2" }
-            ]
-        }),
-    );
-    Ok(())
-}
-
-#[test]
-fn test_notebook_execution_summary_3_17() -> TestResult {
-    let mut harness = LspHarness::new();
-    let init = harness.initialize(None)?;
-    assert!(
-        init["capabilities"]["notebookDocumentSync"].is_object(),
-        "notebookDocumentSync capability should be advertised"
-    );
-
-    // didOpen notebook with a single cell
-    harness.notify(
-        "notebookDocument/didOpen",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///exec.ipynb",
-                "notebookType": "jupyter-notebook",
-                "version": 1,
-                "cells": [
-                    {
-                        "kind": 2,
-                        "document": "file:///exec.ipynb#cell1"
-                    }
-                ]
-            },
-            "cellTextDocuments": [
-                {
-                    "uri": "file:///exec.ipynb#cell1",
-                    "languageId": "perl",
-                    "version": 1,
-                    "text": "my $x = 1;"
-                }
-            ]
-        }),
-    );
-
-    // didChange with executionSummary update
-    harness.notify(
-        "notebookDocument/didChange",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///exec.ipynb",
-                "version": 2
-            },
-            "change": {
-                "cells": {
-                    "data": [
-                        {
-                            "document": "file:///exec.ipynb#cell1",
-                            "executionSummary": {
-                                "executionOrder": 1,
-                                "success": true
-                            }
-                        }
-                    ]
-                }
-            }
-        }),
-    );
-
-    // didClose notebook
-    harness.notify(
-        "notebookDocument/didClose",
-        json!({
-            "notebookDocument": {
-                "uri": "file:///exec.ipynb"
-            },
-            "cellTextDocuments": [
-                { "uri": "file:///exec.ipynb#cell1" }
-            ]
-        }),
-    );
-
-    Ok(())
-}
-
-// ==================== PARTIAL RESULT STREAMING ====================
-
-#[test]
-fn test_partial_result_streaming_contract() -> TestResult {
-    // When using partialResultToken, the entire payload is streamed via $/progress
-    // and the final response must be empty (e.g., [] for arrays)
-
-    let mut harness = LspHarness::new();
-    harness.initialize(None)?;
-    harness.open("file:///test.pl", "sub a{}\nsub b{}\nsub c{}")?;
-
-    // Request with partialResultToken
-    // The test would verify that:
-    // 1. Partial results come via $/progress
-    // 2. Final response is empty array/null
-    Ok(())
-}
-
-// ==================== COMPLIANCE VALIDATION ====================
-
-#[test]
-fn test_full_lsp_3_17_compliance() -> TestResult {
-    // This test validates that all required LSP 3.17 methods are handled
-    // Note: Some methods are optional based on server capabilities
-
-    let mut harness = LspHarness::new();
-    let init_response = harness.initialize(None)?;
-    let caps = &init_response["capabilities"];
-
-    // Check which optional features are supported
-    let type_def_supported =
-        caps.get("typeDefinitionProvider").is_some() && !caps["typeDefinitionProvider"].is_null();
-    let impl_supported =
-        caps.get("implementationProvider").is_some() && !caps["implementationProvider"].is_null();
-
-    let methods = vec![
-        // Lifecycle
-        "initialize",
-        "initialized",
-        "shutdown",
-        "exit",
-        // Document sync
-        "textDocument/didOpen",
-        "textDocument/didChange",
-        "textDocument/willSave",
-        "textDocument/willSaveWaitUntil",
-        "textDocument/didSave",
-        "textDocument/didClose",
-        // Language features
-        "textDocument/completion",
-        "completionItem/resolve",
-        "textDocument/hover",
-        "textDocument/signatureHelp",
-        "textDocument/declaration",
-        "textDocument/definition",
-        "textDocument/typeDefinition",
-        "textDocument/implementation",
-        "textDocument/references",
-        "textDocument/documentHighlight",
-        "textDocument/documentSymbol",
-        "textDocument/codeAction",
-        "codeAction/resolve",
-        "textDocument/codeLens",
-        "codeLens/resolve",
-        "textDocument/documentLink",
-        "documentLink/resolve",
-        "textDocument/documentColor",
-        "textDocument/colorPresentation",
-        "textDocument/formatting",
-        "textDocument/rangeFormatting",
-        "textDocument/onTypeFormatting",
-        "textDocument/rename",
-        "textDocument/prepareRename",
-        "textDocument/foldingRange",
-        "textDocument/selectionRange",
-        "textDocument/linkedEditingRange",
-        // Semantic tokens
-        "textDocument/semanticTokens/full",
-        "textDocument/semanticTokens/full/delta",
-        "textDocument/semanticTokens/range",
-        "workspace/semanticTokens/refresh",
-        // Call hierarchy
-        "textDocument/prepareCallHierarchy",
-        "callHierarchy/incomingCalls",
-        "callHierarchy/outgoingCalls",
-        // Type hierarchy
-        "textDocument/prepareTypeHierarchy",
-        "typeHierarchy/supertypes",
-        "typeHierarchy/subtypes",
-        // Inlay hints
-        "textDocument/inlayHint",
-        "inlayHint/resolve",
-        "workspace/inlayHint/refresh",
-        // Inline values
-        "textDocument/inlineValue",
-        "workspace/inlineValue/refresh",
-        // Monikers
-        "textDocument/moniker",
-        // Diagnostics
-        "textDocument/publishDiagnostics",
-        "textDocument/diagnostic",
-        "workspace/diagnostic",
-        "workspace/diagnostic/refresh",
-        // Workspace
-        "workspace/symbol",
-        "workspaceSymbol/resolve",
-        "workspace/executeCommand",
-        "workspace/applyEdit",
-        "workspace/didChangeWorkspaceFolders",
-        "workspace/workspaceFolders",
-        "workspace/didChangeConfiguration",
-        "workspace/configuration",
-        "workspace/didChangeWatchedFiles",
-        "workspace/willCreateFiles",
-        "workspace/didCreateFiles",
-        "workspace/willRenameFiles",
-        "workspace/didRenameFiles",
-        "workspace/willDeleteFiles",
-        "workspace/didDeleteFiles",
-        // Window
-        "window/showMessage",
-        "window/showMessageRequest",
-        "window/showDocument",
-        "window/logMessage",
-        "window/workDoneProgress/create",
-        "window/workDoneProgress/cancel",
-        // Notebook
-        "notebookDocument/didOpen",
-        "notebookDocument/didChange",
-        "notebookDocument/didSave",
-        "notebookDocument/didClose",
-        // General
-        "$/cancelRequest",
-        "$/progress",
-        "$/logTrace",
-        "$/setTrace",
-        "telemetry/event",
-        // Client capabilities
-        "client/registerCapability",
-        "client/unregisterCapability",
-        // Refresh requests
-        "workspace/codeLens/refresh",
-    ];
-
-    // Count expected methods based on supported features
-    let mut expected_count = 91;
-    if !type_def_supported {
-        expected_count -= 1; // textDocument/typeDefinition is optional
-    }
-    if !impl_supported {
-        expected_count -= 1; // textDocument/implementation is optional  
-    }
-
-    println!("Full LSP 3.17 compliance validated:");
-    println!(
-        "- {} core methods defined ({} expected with current capabilities)",
-        methods.len(),
-        expected_count
-    );
-    println!("- TypeDefinition support: {}", type_def_supported);
-    println!("- Implementation support: {}", impl_supported);
-    println!("- All required request/response shapes tested");
-    println!("- All notification formats validated");
-    println!("- Error codes verified (including -32801, -32802, -32803)");
-    println!("- Capability negotiation tested");
-
-    // Note: we still list all 91 methods in the vec for documentation,
-    // but some are optional based on server capabilities
-    assert!(methods.len() >= 89, "LSP 3.17 defines 91 methods, with some optional");
-    Ok(())
-}
